backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
yfinance==0.2.36
pandas==1.3.3
requests==2.26.0
python-dotenv==0.19.0
aiohttp==3.8.1
websockets==10.1
redis==4.3.4
prometheus-client==0.14.1
gunicorn==20.1.0
tenacity==8.0.1
cachetools==5.0.0
python-dateutil==2.8.2
flask-limiter==2.6.3
flask-socketio==5.3.2
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
pytest-mock==3.10.0 